Remove dummy return values for returns from void methods
As per preliminary tests (and the documentation), the dummy values are
only to be used in intermediate states of the firm representation and
cannot be used for code generation.

Instead, the void methods now have an empty result list, the projection
nodes for using the values have been removed entirely.