Fix latent bug in Phi code generation with Cond predecessors
Previously, the predecessor values of the Phi node and the Projections
from the Cond the Phi node refers to were not assigned to each other
correctly:

It is not sufficient to check the number of the predecessor of the Phi
node. Instead, the number of the predecessor of the block needs to be
compared to the Proj Nr attribute of the predecessor Projs from the Cond
node.